+++ ProgAvanceeProjectWPF/DAO/RideDAO.cs
@@ -108,7 +108,7 @@ internal class RideDAO : DAO<Ride>
                             reader.GetInt32("numRide"),
                             reader.GetString("placeDeparture"),
                             reader.GetDateTime("dateDeparture"),
-                            reader.GetFloat("packageFee"),
+                            reader.GetDouble("packageFee"),
                             categoryDAO.Find(numCategory)
                             );
                         rides.Add(ride);
@@ -145,7 +145,7 @@ internal class RideDAO : DAO<Ride>
                             reader.GetInt32("num"),
                             reader.GetString("placeDeparture"),
                             reader.GetDateTime("dateDeparture"),
-                            reader.GetFloat("packageFee"),
+                            reader.GetDouble("packageFee"),
                             categoryDAO.Find(reader.GetInt32("numCategory"))
                             
                         );
